berry_blender: migrate background resources
Move the verified center and outer palettes, center tilemap, and three
background streams to the Berry Blender resource hierarchy. Keep the JP
compressed bytes intact, retain the unclassified 0x1F1A data suffix, and
preserve all six historical resource addresses through linker aliases.

Verified: clean make -j4, make compare, both complete ROM spans, extracted
resource bytes, decompressed upstream graphics equivalence, map/symbol
addresses, legacy aliases, and the Berry Blender resource audit.

Progress: strict C 8597/16023 (53.65%); source inventory 8385/8452 (99.21%);
top-level path coverage 298/310 (96.13%); transitional files 792; JP INCBIN
8580; raw ROM ASM 2540 references / 6808489 bytes.

+++ include/graphics.h
@@ -4830,7 +4830,8 @@ extern const u16 gBerryBlenderMiscPalette[];
 extern const u16 gBerryBlenderArrowPalette[];
 extern const u32 gBerryBlenderCenter_Gfx[];
 extern const u32 gBerryBlenderOuter_Gfx[];
-extern const u32 gBerryBlenderOuter_Tilemap[];
+// JP's byte-identical compressed stream is 0x311 bytes, not u32-aligned.
+extern const u8 gBerryBlenderOuter_Tilemap[];
 
 // Slot Machine
 extern const u32 gSlotMachineDigitalDisplay_Gfx[];
